Conflicts:
	app/controllers/api/v1/appliances_controller.rb
	app/models/appliance.rb
	spec/services/optimizer_spec.rb

+++ app/models/appliance.rb
@@ -11,6 +11,9 @@
 #  name                                :string(255)
 #  created_at                          :datetime
 #  updated_at                          :datetime
+#  fund_id                             :integer
+#  last_billing                        :datetime
+#  state_explanation                   :string(255)
 #
 
 class Appliance < ActiveRecord::Base
@@ -74,13 +77,21 @@ class Appliance < ActiveRecord::Base
     BillingService::bill_appliance(self, Time.now, "Final billing action prior to appliance destruction.",false)
   end
 
-  def create_dev_mode_property_set
-    self.dev_mode_property_set = DevModePropertySet.create_from(appliance_type) unless self.dev_mode_property_set
+  def create_dev_mode_property_set(options={})
+    unless self.dev_mode_property_set
+      set = DevModePropertySet.create_from(appliance_type)
+      set.preference_memory = options[:preference_memory] if options[:preference_memory]
+      set.preference_cpu = options[:preference_cpu] if options[:preference_cpu]
+      set.preference_disk = options[:preference_disk] if options[:preference_disk]
+      self.dev_mode_property_set = set
+      set.appliance = self
+    end
   end
 
+  private
+
   def remove_appliance_configuration_instance_if_needed
-    if appliance_configuration_instance.appliances.blank?
-      Air.action_logger.info "         -- appliance configuration intance #{appliance_configuration_instance.id} destroyed by appliance #{id}"
+    if appliance_configuration_instance && appliance_configuration_instance.appliances.blank?
       appliance_configuration_instance.destroy
     end
   end
